Refactor directory structure to be clearer
