+++ frontend/src/components/pages/ComponentLibrary.tsx
@@ -11,8 +11,8 @@ import CorrectedMultiSelect from "../teacher/results/StudentAnswersSection/quest
 import CorrectedShortAnswer from "../teacher/results/StudentAnswersSection/question-elements/CorrectedShortAnswer";
 import StudentList from "../teacher/results/StudentList";
 import AddStudentModal from "../teacher/student-management/AddStudentModal";
-import AddClassroomModal from "../teacher/student-management/classrooms/AddClassroomModal";
-import ClassroomCard from "../teacher/student-management/classrooms/ClassroomCard";
+import AddClassroomModal from "../teacher/student-management/classroom-summary/AddClassroomModal";
+import ClassroomCard from "../teacher/student-management/classroom-summary/ClassroomCard";
 
 const defaultValues = {
   className: "",
