+++ frontend/src/components/pages/teacher/ClassroomsPage.tsx
@@ -26,8 +26,8 @@ import LoadingState from "../../common/info/LoadingState";
 import EmptyClassroomsMessage from "../../common/info/messages/EmptyClassroomsMessage";
 import Pagination from "../../common/table/Pagination";
 import usePaginatedData from "../../common/table/usePaginatedData";
-import AddClassroomModal from "../../teacher/student-management/classrooms/AddClassroomModal";
-import ClassroomCard from "../../teacher/student-management/classrooms/ClassroomCard";
+import AddClassroomModal from "../../teacher/student-management/classroom-summary/AddClassroomModal";
+import ClassroomCard from "../../teacher/student-management/classroom-summary/ClassroomCard";
 
 const ClassroomsPage = (): React.ReactElement => {
   const [tabIndex, setTabIndex] = React.useState<TabEnumClassroom>(
