+++ frontend/src/components/pages/teacher/DisplayClassroomPage/index.tsx
@@ -15,7 +15,7 @@ import HeaderWithButton from "../../../common/HeaderWithButton";
 import FormBreadcrumb from "../../../common/navigation/FormBreadcrumb";
 import RouterTabs from "../../../common/navigation/RouterTabs";
 import AddStudentModal from "../../../teacher/student-management/AddStudentModal";
-import AddClassroomOrStudentPopover from "../../../teacher/student-management/classrooms/AddClassroomOrStudentPopover";
+import AddClassroomOrStudentPopover from "../../../teacher/student-management/view-students/AddClassroomOrStudentPopover";
 import NotFound from "../../NotFound";
 
 import DisplayClassroomAssessmentsPage from "./DisplayClassroomAssessmentsPage";
