+++ frontend/src/components/teacher/student-management/classroom-summary/ClassroomCard.tsx
