+++ frontend/src/components/teacher/student-management/classroom-summary/ClassroomPopover.tsx
